Add src/vendor_loader to load vendor assets
The vendor.js file adds dynamically the asset scripts to the document,
which by default instructs the browser to load them asynchronously. Thus,
there are no guarantees in what order the browser will evaluate them.
To resolve the dependencies among the scripts, and namely to initialize
the `easyXDM.Socket` after the easyXDM module is loaded, we have to use
callback functions when `script.onload` event fires. Also, to tackle with
IE<10 which do not support the `script.onload` event we have to use the
`script.onreadystatechange` event.