Include Runnable in ActionsManager and inject Session
Incompatibilities
-----------------

* ActionsManager no longer instantiates a Session and it's only run
after we obtain a Session which is injected as a constructor argument.

* ActionsManager is no longer "thenable" and it does not hold a promise.
There is no foreseeable use-case for waiting all beacons to be reported
(using a promise). ActionsManager#then is not used anywhere in the codebase.

* ActionsManager no longer replaces `window.sa` with its `run` upon initialization.
Since we want the PluginManager to be as well able to run commands whenever they're declared,
This is going to be a responsibility of Analytics.