Store the window parent document in a variable
When clicking on why or save the order, the `parent window` of the
plugin iframe is null. However the `window` property is being used
numerous times in previous methods and it is not null!

It seems that this happens only in Google Chrome
Version 51.0.2704.103 (64-bit).

Perhaps it's a chrome bug, but it's not fully reproducible as it
happens only in specific sites.

So our workaround is to store a reference of window.parent.document to
an instance variable.