Add Settings.command_queue_name
Define {Skroutz, Alve, Scrooge}AnalyticsObject as namespace.

We will use this setting to initialize a global 'sa' function, or a function with
a name specified by the integrator.

In order to respect the function name specified by an integrator, the
Analytics Tracking Script needs to be changed as below:

<script type="text/javascript">
  (function(a,b,c,d,e,f,g){a['SkroutzAnalyticsObject']=e;a[e]= a[e] || function(){
    (a[e].q = a[e].q || []).push(arguments);};
    f=b.createElement(c);f.async=true;
    f.src=d;g=b.getElementsByTagName(c)[0];g.parentNode.insertBefore(f,g);
  })(window,document,'script','https://analytics.skroutz.gr/analytics.min.js','sa');

  sa('session', 'connect', 'SA-XXXX-XXXX');  // Connect your Account.
</script>

Note:
The only addition to Analytics Tracking Script is: a['SkroutzAnalyticsObject']=e;